+++ experimental/packages/opentelemetry-sdk-node/src/index.ts
@@ -14,8 +14,16 @@ export * as contextBase from '@opentelemetry/api';
 export * as core from '@opentelemetry/core';
 export * as logs from '@opentelemetry/sdk-logs';
 export * as metrics from '@opentelemetry/sdk-metrics';
-export * as node from '@opentelemetry/sdk-trace-node';
 export * as resources from '@opentelemetry/resources';
+export * as trace from '@opentelemetry/sdk-trace';
+
+/**
+ * @deprecated Use `trace` re-export from `@opentelemetry/sdk-trace` instead.
+ */
+export * as node from '@opentelemetry/sdk-trace-node';
+/**
+ * @deprecated Use `trace` re-export from `@opentelemetry/sdk-trace` instead.
+ */
 export * as tracing from '@opentelemetry/sdk-trace-base';
 /* eslint-enable no-restricted-syntax */